Added new configurations

+++ SyncAzureBlob/Program.cs
@@ -10,16 +10,44 @@ namespace SyncAzureBlob
     {
         static void Main(string[] args)
         {
+            bool isDeleteRequired = Convert.ToBoolean("IsDeleteRequired".GetSettings());
             string storeToPath = "StoreIn".GetSettings();
-            string[] runInterval = "RunInterval".GetSettings().Split(':');
-            while(true)
+            string runInterval = string.Empty;
+            try
             {
-                ProcessBlobFiles(storeToPath).Wait();
-                Task.Delay(new TimeSpan(int.Parse(runInterval[0]), int.Parse(runInterval[1]), int.Parse(runInterval[2]), int.Parse(runInterval[3]))).Wait();
+                runInterval = "RunInterval".GetSettings();
+            }
+            catch(Exception)
+            {
+                Console.WriteLine("Couldn't find RunInterval in config. So, the program will run only once.");
+            }
+            string containerNames = "ContainerNames".GetSettings();
+            if (string.IsNullOrEmpty(runInterval) ||
+                (!string.IsNullOrEmpty(runInterval) && "0:0:0:0".Equals(runInterval)))
+            {
+                ProcessBlobFilesWithinContainer(storeToPath, containerNames, isDeleteRequired).Wait();
+            }
+            else
+            {
+                string[] runIntervals = runInterval.Split(':');
+                while (true)
+                {
+                    ProcessBlobFilesWithinContainer(storeToPath, containerNames, isDeleteRequired).Wait();
+                    Task.Delay(new TimeSpan(int.Parse(runIntervals[0]), int.Parse(runIntervals[1]), int.Parse(runIntervals[2]), int.Parse(runIntervals[3]))).Wait();
+                }
             }
         }
 
-        static async Task ProcessBlobFiles(string storeToPath)
+        static async Task ProcessBlobFilesWithinContainer(string storeToPath, string containerNames, bool isDeleteRequired)
+        {
+            var containers = containerNames.Split(':');
+            foreach(var container in containers)
+            {
+                await ProcessBlobFiles(storeToPath, containerNames, isDeleteRequired);
+            }
+        }
+
+        static async Task ProcessBlobFiles(string storeToPath, string containerName, bool isDeleteRequired)
         {
             try
             {
@@ -31,13 +59,13 @@ namespace SyncAzureBlob
                 CloudBlobClient blobClient = storageAccount.CreateCloudBlobClient();
 
                 // Retrieve a reference to a container.
-                CloudBlobContainer container = blobClient.GetContainerReference("logs");
+                CloudBlobContainer container = blobClient.GetContainerReference(containerName);
 
                 // Probably this is not required
                 //container.CreateIfNotExists();
-
+                
                 // Loop over items within the container and output the length and URI.
-                foreach (IListBlobItem item in container.ListBlobs(null, false))
+                foreach (IListBlobItem item in container.ListBlobs(null, useFlatBlobListing: true))
                 {
                     if (item.GetType() == typeof(CloudBlockBlob))
                     {
@@ -49,28 +77,43 @@ namespace SyncAzureBlob
                         CloudBlockBlob blockBlob = container.GetBlockBlobReference(blob.Name);
 
                         // Save blob contents to a file.
-                        string toSave = storeToPath + blob.Name;
+                        string toSave = storeToPath + blob.Name.Replace('/', '\\');
+                        string path = toSave.Substring(0, toSave.LastIndexOf('\\'));
+
+                        bool exists = System.IO.Directory.Exists(path);
+                        if (!exists)
+                            System.IO.Directory.CreateDirectory(path);
                         using (var fileStream = System.IO.File.OpenWrite(toSave))
                         {
                             await blockBlob.DownloadToStreamAsync(fileStream);
                         }
                         Console.WriteLine("Download completed for: {0}", blob.Name);
-                        await blob.DeleteAsync();
-                        Console.WriteLine("Download completed for: {0}", blob.Name);
+                        if (isDeleteRequired)
+                        {
+                            await blob.DeleteAsync();
+                            Console.WriteLine("Deleted: {0}", blob.Name);
+                        }
                     }
-                    else if (item.GetType() == typeof(CloudPageBlob))
-                    {
-                        CloudPageBlob pageBlob = (CloudPageBlob)item;
+                    //else if (item.GetType() == typeof(CloudPageBlob))
+                    //{
+                    //    CloudPageBlob pageBlob = (CloudPageBlob)item;
 
-                        Console.WriteLine("Page blob of length {0}: {1}", pageBlob.Properties.Length, pageBlob.Uri);
+                    //    Console.WriteLine("Page blob of length {0}: {1}", pageBlob.Properties.Length, pageBlob.Uri);
 
-                    }
-                    else if (item.GetType() == typeof(CloudBlobDirectory))
-                    {
-                        CloudBlobDirectory directory = (CloudBlobDirectory)item;
+                    //}
+                    //else if (item.GetType() == typeof(CloudBlobDirectory))
+                    //{
+                    //    CloudBlobDirectory directory = (CloudBlobDirectory)item;
 
-                        Console.WriteLine("Directory: {0}", directory.Uri);
-                    }
+                    //    var subPath = storeToPath + directory.Prefix.Substring(0, directory.Prefix.Length - 1).Replace('/', '\\');
+                    //    bool exists = System.IO.Directory.Exists(subPath);
+
+                    //    if (!exists)
+                    //        System.IO.Directory.CreateDirectory(subPath);
+
+                    //    Console.WriteLine("Directory: {0}", directory.Uri);
+                    //}
+                    Console.WriteLine("{0}{0}", Environment.NewLine);
                 }
             }
             catch (Exception ex)
